removed some DEBUG messages and added new test for checking if dot file exists and then copying it to cache

+++ src/dotgraph.cpp
@@ -132,7 +132,9 @@ void copyToCache(QCString const& absFileName, QCString const& cacheName)
 
   QFile f(cacheName);
   if (f.exists()) f.remove();
-  copyFile(absFileName, Doxygen::dotCacheDir + cacheName);
+
+  QFile g(absFileName);
+  if (g.exists()) copyFile(absFileName, Doxygen::dotCacheDir + cacheName);
 }
 
 //--------------------------------------------------------------------
